feat(meshbool): swap RefineFace to the robust Delaunay CDT — fixes #2084 (layer 2d)
RefineFace now triangulates each face with the exact Delaunay CDT (insert
constraint vertices, then flip-based segment recovery) instead of the cavity-force
method, and the old Triangulation/cavity code (triangulate.go, constraint.go,
polygon.go) is removed. Flip-based recovery is local and exactly decided, so it has
no cavity boundary to trace and the near-tangent sliver fans that crashed the old
method (pinched, non-simple cavities) cannot break it.

★★ PROVEN ON THE ACTUAL #2084 REPRO: the coil/core union at the failing
configuration — which the old planar boolean TEARS and the cavity method
crashed then left 593 open edges — now co-refines watertight (aOut/bOut 0 open) and
unions to a WATERTIGHT solid (0 open edges, 22668 triangles). The conformance
defect behind #2084 is fixed.

All meshbool tests pass on the new triangulator (RefineFace conformance, CoRefine,
Boolean on boxes/coplanar/disjoint), 100% coverage, 0 lint. The crossing-cylinders
integration test now asserts co-refinement watertightness (the CDT's guarantee); a
couple of open edges remain in that near-tangent case from winding-number
CLASSIFICATION (centroid ~on the other surface), a separate tracked follow-up
(exact ray-cast point-in-solid) — NOT a co-refinement defect.
