+++ kernel/meshbool/refine.go
@@ -2,35 +2,32 @@
 
 package meshbool
 
-import (
-	"math/big"
-	"sort"
-)
-
 // RefineFace triangulates a face in conformance with a set of constraint segments
 // — the segments where the other operand's faces cross this one (from
 // IntersectTriangles). Every segment endpoint and every pairwise crossing is
-// inserted as a vertex, then each segment is forced as a chain of edges, so the
-// result subdivides the face along the intersection curve with no segment left
-// crossed by a triangle. This is the per-face core of co-refinement.
+// inserted as a vertex of a Delaunay triangulation, then each segment is forced as
+// a chain of edges by exact flip-based recovery, so the result subdivides the face
+// along the intersection curve with no segment left crossed by a triangle. The
+// Delaunay CDT replaces the old cavity-force method, which #2084's near-tangent
+// sliver fans defeat (pinched, non-simple cavities).
 //
 // PRECONDITION: each segment lies in the face's plane with endpoints on or inside
 // the face (as IntersectTriangles guarantees). Points keep exact coordinates; the
 // caller rounds the returned triangles.
 func RefineFace(face [3]Point, segments [][2]Point) [][3]Point {
-	tr := NewTriangulation(face)
-	for _, v := range constraintVertices(segments, tr.axis) {
-		tr.InsertPoint(v)
+	d := newDelaunayInTriangle(face)
+	for _, v := range constraintVertices(segments, d.axis) {
+		d.Insert(v)
 	}
 	for _, s := range segments {
-		tr.forceSegment(s[0], s[1])
+		d.forceSegment(s[0], s[1])
 	}
-	tris := tr.Triangles()
-	// NewTriangulation stores CCW-in-projection; if the input face was CW there, the
-	// sub-triangles carry the opposite normal. Restore the operand's face orientation
-	// so downstream winding-number classification and the result mesh see the true
-	// outward normal.
-	if orient2(face[0], face[1], face[2], tr.axis) < 0 {
+	tris := d.triangles()
+	// newDelaunayInTriangle stores CCW-in-projection; if the input face was CW there,
+	// the sub-triangles carry the opposite normal. Restore the operand's face
+	// orientation so downstream winding-number classification and the result mesh see
+	// the true outward normal.
+	if orient2(face[0], face[1], face[2], d.axis) < 0 {
 		for i := range tris {
 			tris[i][1], tris[i][2] = tris[i][2], tris[i][1]
 		}
@@ -53,44 +50,4 @@ func constraintVertices(segments [][2]Point, axis int) []Point {
 		}
 	}
 	return out
-}
-
-// forceSegment forces the whole segment [a,b] as triangulation edges, splitting it
-// at every vertex that lies on it so each consecutive sub-edge satisfies ForceEdge's
-// no-vertex-strictly-between precondition.
-func (tr *Triangulation) forceSegment(a, b Point) {
-	if a.Equal(b) {
-		return // degenerate (a Touching point): its vertex is inserted, no edge to force
-	}
-	on := tr.verticesOnSegment(a, b)
-	for k := 0; k+1 < len(on); k++ {
-		tr.ForceEdge(on[k], on[k+1])
-	}
-}
-
-// verticesOnSegment returns the indices of all vertices lying on the closed
-// segment [a,b], ordered from a to b by their exact parameter.
-func (tr *Triangulation) verticesOnSegment(a, b Point) []int {
-	lenSq := segParam(a, b, b) // (b-a)·(b-a)
-	type onVert struct {
-		idx int
-		t   *big.Rat
-	}
-	var on []onVert
-	for i, v := range tr.verts {
-		if !rcollinear(a, b, v) {
-			continue
-		}
-		t := segParam(a, b, v)
-		if t.Sign() < 0 || t.Cmp(lenSq) > 0 {
-			continue // beyond an endpoint
-		}
-		on = append(on, onVert{i, t})
-	}
-	sort.Slice(on, func(i, j int) bool { return on[i].t.Cmp(on[j].t) < 0 })
-	idx := make([]int, len(on))
-	for i, e := range on {
-		idx[i] = e.idx
-	}
-	return idx
 }
